fix: Fix updating of notice messages
Previously, the notice parameter would be discarded in Chatbot.update,
and also the type would always be set to "text" in the "new_content"
part of the message body. This meant that updating a notice would make
it a regular message always.

Both issues are now fixed, and a bit of debug info is added.